feat: bysize page로 이동 및 query 추가
